+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ4V2WREM1HBZSCDHPTZPWTT"
+        "01KJ502Q14YJBDATAM644DGBME"
     )
     r"""The unique identifier of the data source"""