+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ4V2WRJ9S0SZPGBPZ1FWZ9T"
+        "01KJ502Q1F7DSY8F0J6VJETNGH"
     )
     r"""The unique identifier of the data source"""