+++ packages/orq-rc/RELEASES.md
@@ -5959,3 +5959,13 @@ Based on:
 - [python v4.4.0-rc.28] packages/orq-rc
 ### Releases
 - [PyPI v4.4.0-rc.28] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.28 - packages/orq-rc
+
+## 2026-02-23 10:20:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.724.1 (2.836.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.4.0-rc.29] packages/orq-rc
+### Releases
+- [PyPI v4.4.0-rc.29] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.29 - packages/orq-rc